backend.local: Refactor conversion code
